Moved logic to pull products from merchants data from component to reducer;
